+++ src/es6/navigator.js
@@ -207,6 +207,7 @@ class Navigator {
                     that.model.view.markSelectedDay();
                 }
                 that.model.options.dayPicker.onSelect(thisUnix);
+                that.model.options.onSelect(thisUnix);
             });
         }
 
@@ -232,6 +233,7 @@ class Navigator {
                 that.model.state.setViewDateTime('month', month);
                 that.model.view.render();
                 that.model.options.monthPicker.onSelect(month);
+                that.model.options.onSelect(that.state.selected.unix);
             });
         }
 
@@ -257,6 +259,7 @@ class Navigator {
                 that.model.state.setViewDateTime('year', year);
                 that.model.view.render();
                 that.model.options.yearPicker.onSelect(year);
+                that.model.options.onSelect(that.state.selected.unix);
             });
         }
     }